fix: sso loader
